% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_projection_method_functions.R
\name{localProjectionIRF}
\alias{localProjectionIRF}
\title{Estimate local projections}
\usage{
localProjectionIRF(
  data,
  shock,
  target,
  horizons,
  engine = "OLS",
  confidence = 0.1,
  lags = "bic",
  lags_max = 12,
  standardize = FALSE,
  states = NULL,
  NW = FALSE,
  NW_lags = NULL,
  NW_prewhite = NULL,
  IV = FALSE,
  treeBag = 5,
  Cholesky = FALSE,
  cores = NULL,
  seed = NULL
)
}
\arguments{
\item{data}{Dataframe of covariate}

\item{shock}{String denoting variable to shock}

\item{target}{String denoting variable betas to collect}

\item{horizons}{Horizons to forecast out to}

\item{engine}{String declaring how to estimate responses (must be 'OLS', 'AW', or 'OSW')}

\item{confidence}{Alpha used for two-sided confidence intervals}

\item{lags}{Lags to include in regressions. Must be a deterministic number of lag or an information criterion used to chose the number of lags (AIC or BIC). Chosen based on information criteria applied to OLS - future functionality may support tree-based ic.}

\item{lags_max}{Maximum number of lags}

\item{standardize}{Standardize inputs (mean = 0, variance = 1)}

\item{states}{Column name used to define states}

\item{NW}{Newey-West correction on variance-covariance matrix}

\item{NW_lags}{Number of lags to use in Newey-West correction}

\item{NW_prewhite}{Prewhite option for Newey-West correction (see sandwich::NeweyWest function)}

\item{IV}{Perform a two-stage instrumental variable regression (currently only supported for trees)}

\item{treeBag}{Number of tree-based models to bag over}

\item{Cholesky}{Append the initial structural shock from a cholesky decomposition [primarly for replication purposes]}

\item{cores}{Number of cores available to use (must be integer) set -1 to indicate use all available resources}

\item{seed}{Random seed to be used (must be integer)}
}
\value{
An object of class 'IRF', a list of \code{impulseResponse} (a matrix of impulse responses and standard errors),
\code{engine} (the method used to estimate impulses),
\code{target} (the variable that was responding to a shock),
\code{shock} (the variable that was shocked), and
\code{lags} (the number of lags used in estimation).
}
\description{
The primary function of the treeImpulses package. Calculates local projection style impulse responses using either linear regression
or tree-based estimators. The function supports state-dependent and instrumental variable routines. The user may choose between
trees based on Athey and Wager (2019) [\url{https://arxiv.org/abs/1510.04342}] or Oprescu et al (2019) [\url{http://proceedings.mlr.press/v97/oprescu19a.html}].
}
