% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_projection_chart_functions.R
\name{plot.irf}
\alias{plot.irf}
\title{plot.irf}
\usage{
\method{plot}{irf}(Data)
}
\arguments{
\item{Data}{An object of class IRF to be charted}
}
\description{
Plot a single impulse response
}
