% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_projection_chart_functions.R
\name{plot.irf_var}
\alias{plot.irf_var}
\title{plot.irf_var}
\usage{
\method{plot}{irf_var}(Data, state = NULL)
}
\arguments{
\item{Data}{An object of class IRF_VAR to be charted}

\item{state}{State label used to define groups in state-dependent IRF_var}
}
\description{
Plot a collection of impulse responses
}
